master_bot/requirements.txt
python-telegram-bot>=13.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pika>=1.2.0
